+++ demo/jvmti/hprof/src/hprof_ioname.c
@@ -1,7 +1,7 @@
 /*
- * @(#)hprof_ioname.c	1.10 04/07/27
+ * @(#)hprof_ioname.c	1.12 05/11/17
  * 
- * Copyright (c) 2004 Sun Microsystems, Inc. All Rights Reserved.
+ * Copyright (c) 2006 Sun Microsystems, Inc. All Rights Reserved.
  * 
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are met:
@@ -56,7 +56,7 @@ IoNameIndex
 ioname_find_or_create(const char *name, jboolean *pnew_entry)
 {
     return table_find_or_create_entry(gdata->ioname_table, 
-			(void*)name, strlen(name)+1, pnew_entry, NULL);
+			(void*)name, (int)strlen(name)+1, pnew_entry, NULL);
 }
 
 void
